ref: move guards into field-validation
